Api

LearningToOptimize

LearningToOptimize.FullyConnected

LearningToOptimize.ProblemIterator

LearningToOptimize.Recorder

LearningToOptimize.box_sampler

LearningToOptimize.general_sampler

LearningToOptimize.general_sampler

LearningToOptimize.inconvexhull

LearningToOptimize.line_sampler

LearningToOptimize.load_parameters

LearningToOptimize.load_parameters

LearningToOptimize.make_convex!

LearningToOptimize.make_convex!

LearningToOptimize.record

LearningToOptimize.record

LearningToOptimize.save

LearningToOptimize.scaled_distribution_sampler

LearningToOptimize.solve_and_record

LearningToOptimize.solve_batch

LearningToOptimize.update_model!

LearningToOptimize.update_model!

LearningToOptimize.FullyConnected — Method

FullyConnected(input_size::Int, hidden_sizes::Vector{Int}, output_size::Int)::FullyConnected

Create a fully connected neural network with input_size inputs, hidden_sizes hidden layers and output_size outputs. Adds a pass through layer for each layer.

LearningToOptimize.ProblemIterator — Type

ProblemIterator(ids::Vector{UUID}, pairs::Dict{VariableRef, Vector{Real}})

Iterator for optimization problem instances.

LearningToOptimize.Recorder — Type

Recorder(filename; primal_variables=[], dual_variables=[], filterfn=(model)-> termination_status(model) == MOI.OPTIMAL)

Recorder of optimization problem solutions.

LearningToOptimize.box_sampler — Method

function box_sampler(
    original_parameter::T,
    num_s::F,
    range_p::AbstractVector{T}=0.8:0.01:1.25,
) where {T<:Real,F<:Integer}

Uniformly sample values around the original parameter value over a discrete range inside a box.

LearningToOptimize.general_sampler — Method

general_sampler(
    file::AbstractString;
    samplers::Vector{Function}=[
        (original_parameters) -> scaled_distribution_sampler(original_parameters, 1000),
        LearningToOptimize.line_sampler, 
        (original_parameters) -> box_sampler(original_parameters, 10),
    ],
    batch_id::UUID=uuid1(),
    save_file::AbstractString=split(file, ".mof.json")[1] * "_input_" * string(batch_id),
    filetype::Type{T}=ArrowFile
) where {T<:FileType}

This function is a general sampler that uses a set of samplers to sample the parameter space. It loads the underlying model from file and samples the parameters. It saves the sampled parameters to save_file.

LearningToOptimize.general_sampler — Method

function general_sampler(
    original_parameters::Vector{T};
    samplers::Vector{Function}=[
        (original_parameters) -> scaled_distribution_sampler(original_parameters, 1000),
        LearningToOptimize.line_sampler, 
        (original_parameters) -> box_sampler(original_parameters, 10),
    ]
) where {T<:Real}

This function is a general sampler that uses a set of samplers to sample the parameter space.

LearningToOptimize.inconvexhull — Method

inconvexhull(training_set::Matrix{Float64}, test_set::Matrix{Float64})

Check if new points are inside the convex hull of the given points. Solves a linear programming problem to check if the points are inside the convex hull.

LearningToOptimize.line_sampler — Method

function line_sampler(
    original_parameters::Vector{T},
    parameter_indexes::Vector{F},
    range_p::AbstractVector{T},
) where {T<:Real,F<:Integer}

This sampler returns a set of parameters that for a line in one dimension of the parameter space. The idea is to change the value of one parameter and keep the rest constant.

LearningToOptimize.load_parameters — Method

load_parameters(file::AbstractString)

Load the parameters from a saved jump model.

LearningToOptimize.load_parameters — Method

load_parameters(model::JuMP.Model)

Load the parameters from a JuMP model.

LearningToOptimize.make_convex! — Method

function make_convex!(chain::PairwiseFusion; tol = 1e-6)

Make a PairwiseFusion model convex by making sure all dense layers (a part from the first) have positive weights. This procedure only makes sense for single output fully connected models.

LearningToOptimize.make_convex! — Method

function make_convex!(model::FullyConnected; tol = 1e-6)

Make a FullyConnected model convex by making sure all dense layers (a part from the first) have positive weights. This procedure only makes sense for single output fully connected models.

LearningToOptimize.record — Method

record(recorder::Recorder{ArrowFile}, id::UUID)

Record optimization problem solution to an Arrow file.

LearningToOptimize.record — Method

record(recorder::Recorder{CSVFile}, id::UUID)

Record optimization problem solution to a CSV file.

LearningToOptimize.save — Method

save(problem_iterator::ProblemIterator, filename::AbstractString, file_type::Type{T})

Save optimization problem instances to a file.

LearningToOptimize.scaled_distribution_sampler — Method

function scaled_distribution_sampler(
    original_parameters::Vector{T},
    num_s::F;
    rng::AbstractRNG=Random.GLOBAL_RNG,
    scaler_multiplier::Distribution=Uniform(0.8, 1.25),
    distribution::Distribution=MvLogNormal(fill(-(1.05 .^ 2) ./ 2.0, length(original_parameters)), 1.05)
) where {T<:Real,F<:Integer}

Sample from a distribution and scale the parameters by a random value over a uniform distribution.

LearningToOptimize.solve_and_record — Method

solve_and_record(problem_iterator::ProblemIterator, recorder::Recorder, idx::Integer)

Solve an optimization problem and record the solution.

LearningToOptimize.solve_batch — Method

solve_batch(problem_iterator::AbstractProblemIterator, recorder)

Solve a batch of optimization problems and record the solutions.

LearningToOptimize.update_model! — Method

update_model!(model::JuMP.Model, pairs::Dict, idx::Integer)

Update the values of parameters in a JuMP model.

LearningToOptimize.update_model! — Method

update_model!(model::JuMP.Model, p::VariableRef, val::Real)

Update the value of a parameter in a JuMP model.